# Conflicts:
#	yarn.lock

+++ typescript/db/src/repository/index.ts
@@ -1,63 +1,102 @@
-import { PrismaClient, Prisma } from "@prisma/client";
-
+import { Prisma } from "@prisma/client";
 import { DbLabel, Repository } from "@labelflow/common-resolvers";
 import { Image } from "@labelflow/graphql-types";
+import slugify from "slugify";
+import { prisma } from "./prisma-client";
 import {
   getUploadTargetHttp,
   getFromStorage,
   putInStorage,
   deleteFromStorage,
 } from "./upload-supabase";
-import { countLabels, listLabels } from "./label";
+import {
+  addWorkspace,
+  getWorkspace,
+  listWorkspace,
+  updateWorkspace,
+} from "./workspace";
+import { listLabels, countLabels } from "./label";
 import { castObjectNullsToUndefined } from "./utils";
-
-const prisma = new PrismaClient();
+import {
+  checkUserAccessToDataset,
+  checkUserAccessToImage,
+  checkUserAccessToLabel,
+  checkUserAccessToLabelClass,
+  checkUserAccessToWorkspace,
+} from "./access-control";
 
 export const repository: Repository = {
   image: {
-    add: async (image) => {
-      const createdImage = await prisma.image.create({ data: image });
+    add: async (image, user) => {
+      await checkUserAccessToDataset({ where: { id: image.datasetId }, user });
+      const createdImage = await prisma.image.create({
+        data: castObjectNullsToUndefined(image),
+      });
       return createdImage.id;
     },
-    count: (where) =>
-      prisma.image.count({
-        where: castObjectNullsToUndefined(where),
-      }),
-    getById: async (id) =>
-      (await prisma.image.findUnique({
-        where: { id },
-      })) as unknown as Image,
-
-    list: (where, skip = undefined, first = undefined) =>
-      prisma.image.findMany(
+    count: async (whereWithUser) => {
+      const { user, ...where } = whereWithUser ?? { user: undefined };
+      return await prisma.image.count({
+        where: castObjectNullsToUndefined({
+          ...where,
+          dataset: {
+            workspace: { memberships: { some: { userId: user?.id } } },
+          },
+        }),
+      });
+    },
+    get: async (where, user) => {
+      await checkUserAccessToImage({ where, user });
+      return (await prisma.image.findUnique({
+        where,
+      })) as unknown as Image;
+    },
+    list: async (whereWithUser, skip = undefined, first = undefined) => {
+      const { user, ...where } = whereWithUser ?? { user: undefined };
+      return await prisma.image.findMany(
         castObjectNullsToUndefined({
-          where: castObjectNullsToUndefined(where),
+          where: castObjectNullsToUndefined({
+            ...where,
+            dataset: {
+              workspace: { memberships: { some: { userId: user?.id } } },
+            },
+          }),
           orderBy: { createdAt: Prisma.SortOrder.asc },
           skip,
           take: first,
         })
-      ),
+      );
+    },
+    delete: async (where, user) => {
+      await checkUserAccessToImage({ where, user });
+      await prisma.image.delete({ where });
+    },
   },
   label: {
-    add: async (label) => {
+    add: async (label, user) => {
+      await checkUserAccessToImage({ where: { id: label.imageId }, user });
       const createdLabel = await prisma.label.create({ data: label });
       return createdLabel.id;
     },
     count: countLabels,
-    delete: async (id) => {
+    delete: async ({ id }, user) => {
+      await checkUserAccessToLabel({ where: { id }, user });
       await prisma.label.delete({ where: { id } });
     },
     /* Needs to be casted as Prisma doesn't let us specify
      * the type for geometry */
-    getById: (id) =>
-      prisma.label.findUnique({
-        where: { id },
-      }) as unknown as Promise<DbLabel>,
-    update: async (id, label) => {
+    get: async (where, user) => {
+      await checkUserAccessToLabel({ where, user });
+      return (await prisma.label.findUnique({
+        where,
+      })) as unknown as DbLabel;
+    },
+    update: async (where, label, user) => {
+      await checkUserAccessToLabel({ where, user });
       try {
         if (label) {
           await prisma.label.update({
-            where: { id },
+            where,
             data: castObjectNullsToUndefined(label),
           });
         }
@@ -69,34 +108,56 @@ export const repository: Repository = {
     list: listLabels,
   },
   labelClass: {
-    add: async (labelClass) => {
+    add: async (labelClass, user) => {
+      await checkUserAccessToDataset({
+        where: { id: labelClass.datasetId },
+        user,
+      });
       const createdLabelClass = await prisma.labelClass.create({
-        data: labelClass,
+        data: castObjectNullsToUndefined(labelClass),
       });
       return createdLabelClass.id;
     },
-    count: async (where) =>
-      await prisma.labelClass.count({
-        where: castObjectNullsToUndefined(where),
-      }),
-    delete: async (id) => {
+    count: async (whereWithUser) => {
+      const { user, ...where } = whereWithUser ?? { user: undefined };
+      return await prisma.labelClass.count({
+        where: castObjectNullsToUndefined({
+          ...where,
+          dataset: {
+            workspace: { memberships: { some: { userId: user?.id } } },
+          },
+        }),
+      });
+    },
+    delete: async ({ id }, user) => {
+      await checkUserAccessToLabelClass({ where: { id }, user });
       await prisma.labelClass.delete({ where: { id } });
     },
-    getById: (id) =>
-      prisma.labelClass.findUnique({
-        where: { id },
-      }),
-    list: (where, skip = undefined, first = undefined) =>
-      prisma.labelClass.findMany(
+    get: async (where, user) => {
+      await checkUserAccessToLabelClass({ where, user });
+      return await prisma.labelClass.findUnique({
+        where,
+      });
+    },
+    list: async (whereWithUser, skip = undefined, first = undefined) => {
+      const { user, ...where } = whereWithUser ?? { user: undefined };
+      return await prisma.labelClass.findMany(
         castObjectNullsToUndefined({
-          where: castObjectNullsToUndefined(where),
+          where: castObjectNullsToUndefined({
+            ...where,
+            dataset: {
+              workspace: { memberships: { some: { userId: user?.id } } },
+            },
+          }),
           orderBy: { index: Prisma.SortOrder.asc },
           skip,
           take: first,
         })
-      ),
-    update: async (id, labelClass) => {
+      );
+    },
+    update: async ({ id }, labelClass, user) => {
       try {
+        await checkUserAccessToLabelClass({ where: { id }, user });
         await prisma.labelClass.update({
           where: { id },
           data: castObjectNullsToUndefined(labelClass),
@@ -108,26 +169,65 @@ export const repository: Repository = {
     },
   },
   dataset: {
-    add: async (dataset) => {
-      const createdDataset = await prisma.dataset.create({ data: dataset });
+    add: async ({ workspaceSlug, ...dataset }, user) => {
+      await checkUserAccessToWorkspace({
+        where: { slug: workspaceSlug },
+        user,
+      });
+      const slug = slugify(dataset.name, { lower: true });
+      const createdDataset = await prisma.dataset.create({
+        data: castObjectNullsToUndefined({
+          ...dataset,
+          slug,
+          workspace: {
+            connect: { slug: workspaceSlug },
+          },
+        }),
+      });
       return createdDataset.id;
     },
-    delete: async (id) => {
-      await prisma.dataset.delete({ where: { id } });
+    delete: async (where, user) => {
+      if (
+        (where.id == null && where.slugs == null) ||
+        (where.id != null && where.slugs != null)
+      ) {
+        throw new Error(
+          "You should either specify the id or the slugs when deleting a dataset"
+        );
+      }
+      await checkUserAccessToDataset({ where, user });
+      await prisma.dataset.delete({
+        where: castObjectNullsToUndefined(where),
+      });
     },
-    getById: (id) => {
-      return prisma.dataset.findUnique({ where: { id } });
+    get: async (where, user) => {
+      if (
+        (where.id == null && where.slugs == null) ||
+        (where.id != null && where.slugs != null)
+      ) {
+        throw new Error(
+          "You should either specify the id or the slugs when looking for a dataset"
+        );
+      }
+      await checkUserAccessToDataset({ where, user });
+      const dataset = await prisma.dataset.findUnique({
+        where: castObjectNullsToUndefined(where),
+      });
+      return dataset;
     },
-    getByName: (name) => {
-      return prisma.dataset.findUnique({ where: { name } });
-    },
-    getBySlug: (slug) => {
-      return prisma.dataset.findUnique({ where: { slug } });
-    },
-    update: async (id, dataset) => {
+    update: async (where, dataset, user) => {
+      if (
+        (where.id == null && where.slugs == null) ||
+        (where.id != null && where.slugs != null)
+      ) {
+        throw new Error(
+          "You should either specify the id or the slugs when updating a dataset"
+        );
+      }
+      await checkUserAccessToDataset({ where, user });
       try {
         await prisma.dataset.update({
-          where: { id },
+          where: castObjectNullsToUndefined(where),
           data: castObjectNullsToUndefined(dataset),
         });
         return true;
@@ -135,14 +235,26 @@ export const repository: Repository = {
         return false;
       }
     },
-    list: (_where, skip = undefined, first = undefined) =>
+    list: (where, skip = undefined, first = undefined) =>
       prisma.dataset.findMany(
         castObjectNullsToUndefined({
           orderBy: { createdAt: Prisma.SortOrder.asc },
           skip,
           take: first,
+          where: {
+            workspace: {
+              slug: where?.workspaceSlug,
+              memberships: { some: { userId: where?.user?.id } },
+            },
+          },
         })
       ),
+  },
+  workspace: {
+    add: addWorkspace,
+    get: getWorkspace,
+    list: listWorkspace,
+    update: updateWorkspace,
   },
   upload: {
     delete: deleteFromStorage,
